ServerStuff/requirements.txt
python-multipart
azure-identity[aio]
azure-storage-blob[aio]==12.27.1
fastapi
uvicorn[standard]
gunicorn